new lexer and tests

+++ src/main/c/include/conditions.h
@@ -49,6 +49,6 @@ extern const void * SENTINEL;
 
 
 bool is_null(const void *first, ...);
-bool is_false(int_fast8_t first, ...);
+bool is_false(int first, ...);
